requirements.txt
requests
beautifulsoup4
pandas
sqlalchemy
pymongo
dnspython
playwright
python-dotenv
